+++ aspnet/RVTR.Account.ObjectModel/Models/PaymentModel.cs
@@ -1,26 +1,22 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
 namespace RVTR.Account.ObjectModel.Models
 {
   /// <summary>
-  /// Represents the _Account_ model
+  /// Represents the _Payment_ model
   /// </summary>
-  public class AccountModel : IValidatableObject
+  public class PaymentModel : IValidatableObject
   {
     public int Id { get; set; }
 
-    public AddressModel Address { get; set; }
+    public BankCardModel BankCard { get; set; }
 
     [Required]
     public string Name { get; set; }
 
-    public IEnumerable<PaymentModel> Payments { get; set; }
-
-    public IEnumerable<ProfileModel> Profiles { get; set; }
-
     /// <summary>
-    /// Represents the _Account_ `Validate` method
+    /// Represents the _Payment_ `Validate` method
     /// </summary>
     /// <param name="validationContext"></param>
     /// <returns></returns>