+++ aspnet/RVTR.Account.ObjectModel/Models/ProfileModel.cs
@@ -1,26 +1,26 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
 namespace RVTR.Account.ObjectModel.Models
 {
   /// <summary>
-  /// Represents the _Account_ model
+  /// Represents the _Profile_ model
   /// </summary>
-  public class AccountModel : IValidatableObject
+  public class ProfileModel : IValidatableObject
   {
     public int Id { get; set; }
 
-    public AddressModel Address { get; set; }
+    [EmailAddress, Required]
+    public string Email { get; set; }
 
     [Required]
-    public string Name { get; set; }
+    public NameModel Name { get; set; }
 
-    public IEnumerable<PaymentModel> Payments { get; set; }
-
-    public IEnumerable<ProfileModel> Profiles { get; set; }
+    [Phone, Required]
+    public string Phone { get; set; }
 
     /// <summary>
-    /// Represents the _Account_ `Validate` method
+    /// Represents the _Profile_ `Validate` method
     /// </summary>
     /// <param name="validationContext"></param>
     /// <returns></returns>